+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 
 # SmarterCSV 1.x Change Log
 
-## 1.5.3 (2022-04-30)
-  * rewrote parsing of quoted fields, use "liberal_parsing" by default
+## 1.5.3 (2022-05-03)
+  * completely rewrote line parser
+  * added methods `SmarterCSV.raw_headers` and `SmarterCSV.headers` to allow easy examination of how the headers are processed.
 
 ## 1.5.2 (2022-04-29)
   * added missing keys to the SmarterCSV::KeyMappingError exception message #189 (thanks to John Dell)